feat: domain layer

+++ src/piaps/domain/value_objects/money.py
@@ -1,6 +1,8 @@
+from decimal import Decimal
+
 from piaps.domain.value_objects.base import ValueObject, value_object
 
 
 @value_object
 class Money(ValueObject):
-    value: float
+    value: Decimal
